test(mcp): drive the whole tool surface against a reserved-page corpus
The owner matrix drives every tool in the live table against a two-OWNER
fixture. Its fixture carries no reserved page, it asserts nothing about
the MCP `_meta` response member, and it drives no federated caller - so a
page reserved against remote reads could have leaked through any of the
surfaces it sweeps and every one of its assertions would still have been
green. This is the second pass over the same enumeration.

A SIBLING file, and the design left that open to be decided on evidence:
the two axes cannot share one fixture. The owner matrix builds its
contexts without a transport reach, which resolves to the narrowest, so
the moment its fixture gained a reserved page every owner assertion would
have had to say which reach it was asking at and its gate-off half would
have been measuring two rules at once.

The recipes are SHARED rather than copied. `tests/helpers/tool-probe-
catalogue.ts` now holds the invocation for every code path of every tool,
and both matrices import it - a second copy of "what is the whole
surface" is the enumeration defect the first matrix was built to remove.
The owner matrix keeps its own classification and its 264 tests stay
green over the move.

Every probe goes through the real JSON-RPC `tools/call` path with a
progress token attached, and asserts over the WHOLE serialised response.
That covers the four channels a marker could ride out on in one check:
`structuredContent`, the rendered `content[].text`, the `message` a throw
produces, and the `_meta` member the progress refusal travels in - which
no prior sweep looked at and which is live on this transport.

Two buckets, both executed, because a blanket "no tool may name a
reserved page" would have been a claim this wave does not deliver - and
the first run proved it, with twenty-eight recipes still naming one. A
tool the registry calls covered must withhold; every other recipe's
verdict is measured in `STILL_NAMED_AT_REMOTE` and asserted in both
directions, so a recipe leaving that list is the wave that covered it and
one joining it is a regression.

One real leak found and fixed on the way: `brain_clusters operation=run`
reported community membership filtered by owner scope alone, so a
community naming a reserved page was reported to a remote caller in full.
Both rules are ANDed now, a community either rule withholds is dropped
whole rather than trimmed - a community of seven reported as four is not
a narrower true finding but a false one - and detection stays vault-wide,
because clustering the visible half of a link graph would produce
different communities per caller and write them over each other.

Named narrowing in that same handler: `removed` is now withheld whenever
ANY rule is live rather than only under an owner scope. A removal names a
cluster note whose id is derived from its seed page's path, so at remote
reach it can name a seed the `communities` list has just withheld, and
the field's own stated reasoning - its subject cannot be resolved - is
what makes withholding the only answer.
